Remove pointless header
